% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armb.R
\name{armb}
\alias{armb}
\title{ARMB nonparametric bootstrap}
\usage{
armb(
  Y,
  X,
  FAMILY,
  NSIM,
  MLE,
  TUNE_STEP,
  TUNE_GAMMA,
  NCORES,
  SEED,
  ARM_CONTROL,
  TUNE_CONTROL,
  VERBOSE = FALSE
)
}
\arguments{
\item{NCORES}{Number of cores used via the \link[pbapply]{pblapply} function.}

\item{SEED}{Argument passed to \code{set.seed()} for reproducibility purposes in bootstrap samples.}

\item{DATA}{Dataset.}

\item{R}{Number of bootstrap samples.}
}
\description{
Fast nonparametric bootstrap for glm parameters using stochastic approximations.
}
\details{
Returns a tibble with classes \code{bootstraps}, \code{rset}, \code{tbl_df}, \code{tbl}, \code{data.frame}.
See \link[rsample]{bootstraps} for further details. The tibble include a column
with bootstrap estimates obtained using \link[sgd]{sgd} algorithms.
}
\references{
Boris T. Polyak and Anatoli B. Juditsky. Acceleration of stochastic
approximation by averaging. \emph{SIAM Journal on Control and Optimization},
30(4):838-855, 1992.
}
