% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armb.R
\name{fastBoot}
\alias{fastBoot}
\title{ARMB nonparametric bootstrap}
\usage{
fastBoot(
  DATA,
  R,
  FORMULA,
  MODEL = "glm",
  MODEL.CONTROL = list(),
  SGD.CONTROL = list(method = "asgd", npasses = 1, pass = TRUE),
  FR = 0.1,
  SEED = 123,
  NCORES = 1,
  TUNEFLAG = T,
  TUNE.CONTROL = list(FR_N = 1, MAX_ATTEMPTS = 10, TOL = 1e-05)
)
}
\arguments{
\item{DATA}{Dataset.}

\item{R}{Number of bootstrap samples.}

\item{SEED}{Argument passed to \code{set.seed()} for reproducibility purposes in bootstrap samples.}

\item{NCORES}{Number of cores used via the \link[pbapply]{pblapply} function.}
}
\description{
Fast nonparametric bootstrap for glm parameters using stochastic approximations.
}
\details{
Returns a tibble with classes \code{bootstraps}, \code{rset}, \code{tbl_df}, \code{tbl}, \code{data.frame}.
See \link[rsample]{bootstraps} for further details. The tibble include a column
with bootstrap estimates obtained using \link[sgd]{sgd} algorithms.
}
\references{
Boris T. Polyak and Anatoli B. Juditsky. Acceleration of stochastic
approximation by averaging. \emph{SIAM Journal on Control and Optimization},
30(4):838-855, 1992.
}
