% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logistic_wrapper}
\alias{logistic_wrapper}
\title{Compute nll and ngr for logistic regression}
\usage{
logistic_wrapper(Y, X, THETA)
}
\arguments{
\item{Y}{response}

\item{X}{design matrix}

\item{THETA}{parameter vector}
}
\value{
Provides a list with the negative log likelihood and its gradient
}
\description{
Used for internal testing
}
