% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resampleN}
\alias{resampleN}
\title{Resample sequence}
\usage{
resampleN(N, SEED)
}
\arguments{
\item{N}{length of the sequence}

\item{SEED}{seed for rng}
}
\description{
Resample with replacement the sequence from 0 to N-1
}
