% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shuffleRows}
\alias{shuffleRows}
\title{Randomly shuffle rows of a matrix}
\usage{
shuffleRows(X, SEED)
}
\arguments{
\item{X}{matrix}

\item{SEED}{seed for rng}
}
\description{
Randomly shuffle rows of a matrix
}
