% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shuffleVec}
\alias{shuffleVec}
\title{Randomly shuffle elements of a vector}
\usage{
shuffleVec(X, SEED)
}
\arguments{
\item{X}{vector}

\item{SEED}{seed for rng}
}
\description{
Randomly shuffle elements of a vector
}
