% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sliceMat}
\alias{sliceMat}
\title{Resample sequence}
\usage{
sliceMat(SLICE_IDX, X)
}
\arguments{
\item{N}{length of the sequence}

\item{SEED}{seed for rng}
}
\description{
Resample with replacement the sequence from 0 to N-1
}
