% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_glm}
\alias{test_glm}
\title{Compute nll and ngr for logistic regression}
\usage{
test_glm(Y, X, FAMILY, LINK, THETA)
}
\arguments{
\item{Y}{response}

\item{X}{design matrix}

\item{THETA}{parameter vector}
}
\value{
Provides a list with the negative log likelihood and its gradient
}
\description{
Used for internal testing
}
