% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tune_armLR}
\alias{tune_armLR}
\title{tune logistic regression via ARM}
\usage{
tune_armLR(
  Y,
  X,
  THETA0,
  MAXT,
  BURN,
  BATCH,
  STEPSIZE0,
  SCALE,
  MAXA,
  PAR1,
  PAR2,
  PAR3,
  AUTO_STOP,
  SKIP_PRINT,
  SEED,
  VERBOSE
)
}
\arguments{
\item{Y}{response}

\item{X}{design matrix}

\item{THETA0}{starting values}

\item{MAXT}{max number of iterations}

\item{BURN}{burn-in period}

\item{BATCH}{mini-natch dimension}

\item{STEPSIZE0}{initial stepsize}

\item{SCALE}{multiplying factor}

\item{MAXA}{maximum number of attempts}

\item{PAR1}{par1}

\item{PAR2}{par2}

\item{PAR3}{par3}

\item{SKIP_PRINT}{How many iterations to skip before printing diagnostics}

\item{VERBOSE}{verbose output}

\item{MAXE}{max number of epochs}
}
\description{
tune logistic regression via ARM
}
